+++ frontend/src/__tests__/Header.test.tsx
@@ -1,4 +1,4 @@
-import { describe, it, expect, vi } from 'vitest'
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import Header from '../components/Header'
 
@@ -13,8 +13,15 @@ const defaultProps = {
   refreshNonce: 0,
 }
 
-// ClusterHealthBadge fetches on mount; stub fetch globally so tests don't hit the network.
-vi.stubGlobal('fetch', vi.fn(() => Promise.resolve(new Response(null, { status: 404 }))))
+// ClusterHealthBadge fetches on mount; stub fetch per-test so we don't hit the
+// network and the stub doesn't leak into other test files sharing the global.
+beforeEach(() => {
+  vi.stubGlobal('fetch', vi.fn(() => Promise.resolve(new Response(null, { status: 404 }))))
+})
+
+afterEach(() => {
+  vi.unstubAllGlobals()
+})
 
 describe('Header', () => {
   it('renders the OpenHands logo image', () => {
